// source/src/models/usuario-alta-dto.ts
/* tslint:disable */
export interface UsuarioAltaDto {

  /**
   * Ingrese apellido identificador de usuario
   */
  apellido: string;

  /**
   * Ingrese el correo del usuario
   */
  correo: string;

  /**
   * Ingrese el Cuil del usuario
   */
  cuil: string;

  /**
   * Ingrese Nombre identificador de usuario
   */
  nombre: string;

  /**
   * Ingrese la clave del usuario
   */
  password: string;
}